packages/Server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
email-validator==2.1.0